Added unsigned and float tests (some)

+++ src/resources/pyllars/pyllars.hpp
@@ -293,7 +293,7 @@ public:
 
      static PyMethodDef _methods[];
 
-    std::function<__int128_t()> asDouble;
+    std::function<double()> asDouble;
 
     PyObject* _referenced;
     size_t _depth;
